Apply fixture to reset project context
